captureconvert: add depth intrinsics and extrinsics if available

+++ corvis/scripts/captureconvert.py
@@ -5,6 +5,8 @@ from collections import defaultdict
 import math
 import sys
 import getopt
+import os.path
+import json
 
 camera_type = 1
 accel_type = 20
@@ -121,6 +123,26 @@ print "Wrote", wrote_bytes/1e6, "Mbytes"
 for key in wrote_packets:
     print "Type", key, "-", wrote_packets[key], "packets"
 
-with open(path + "calibration.json", 'r') as f:
-    with open(output_filename + ".json", 'w') as t:
-        t.write(f.read())
+
+
+def transform(W, T, v):
+    from numpy import asarray, cross, eye, matrix
+    from scipy.linalg import expm3
+    return list(expm3(cross(eye(3), asarray(W))).dot(asarray(v)) + asarray(T))
+
+if os.path.isfile(path + "CameraParameters.txt"):
+    with open(path + "CameraParameters.txt", 'r') as p:
+        d_e_c = p.readline().split()
+        with open(path + "calibration.json", 'r') as f:
+            cal = defaultdict(int, json.loads(f.read()))
+            del cal['px']; del cal['py'] # ignored and often uninitialized
+            # g_accel_depth = g_accel_color * (g_color_depth)^-1
+            depth_Tc = transform([cal['Wc0'],cal['Wc1'],cal['Wc2']], [cal['Tc0'],cal['Tc1'],cal['Tc2']], map(lambda x: -x/1000., map(float,d_e_c[6:9])))
+            cal['depth'] = dict(zip("imageWidth imageHeight Fx Fy Cx Cy Wc0 Wc1 Wc2 Tc0 Tc1 Tc2".split(), map(int,d_e_c[:2]) + map(float,d_e_c[2:6]) + [cal['Wc0'],cal['Wc1'],cal['Wc2']] + depth_Tc))
+            with open(output_filename + ".json", 'w') as t:
+                 t.write(json.dumps(cal))
+            print "Added depth intrinsics and extrinsics to " + output_filename + ".json"
+else:
+    with open(path + "calibration.json", 'r') as f:
+        with open(output_filename + ".json", 'w') as t:
+            t.write(f.read())
